pyproject.toml
[tool.poetry]
name = "orders"
version = "0.1.0"
description = ""
authors = ["Aratimius <aratimius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
django-restframework = "^0.0.1"
ipython = "^8.32.0"
black = "^25.1.0"
psycopg2-binary = "^2.9.10"
shell = "^1.0.1"
python-dotenv = "^1.0.1"
coverage = "^7.6.12"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
